URLs now include a locale, /always/.
Fixed a subordinate issue which arised, that the routing stack passes arrays instead of strings through the routing stack. Currently, I probably don't handle this very well, but I'm not sure what to do. Also, Rails 2.3 is going to mess with the routing stack, so I'll just wait and see.