Rails 2.3.0 compat: Generate a 1-line optimisation block.
The Rails commit that added inline comments for eval'd strings questionably
calls generate_optimistaion_block /inside/ the comment as well as out in
the code, requiring that generate_optimisation_block return a string devoid
of newlines to avoid spilling out of the comment and causing a parse error.

The offending line is here:
http://github.com/rails/rails/commit/a2270ef2594b97891994848138614657363f2806#L3R178

Given that routing-filter is a "wild hack," :) it seems better that it adjust
itself to Rails, rather than make allowances in Rails for the overriding of an
arguably "internal" method. Thus, this patch.